Migrate to TypeScript and Jest
