+++ .eslintrc.json
@@ -1,0 +1,41 @@
+{
+  "root": true,
+  "ignorePatterns": ["node_modules", "build", "coverage"],
+  "parser": "@typescript-eslint/parser",
+  "plugins": ["import", "eslint-comments"],
+  "env": { "es6": true },
+  "globals": { "BigInt": true, "console": true, "WebAssembly": true },
+  "overrides": [
+    {
+      "files": ["*.ts"],
+      "parserOptions": {
+        "project": "./tsconfig.json"
+      },
+      "extends": [
+        "eslint:recommended",
+        "plugin:eslint-comments/recommended",
+        "plugin:@typescript-eslint/recommended",
+        "plugin:@typescript-eslint/recommended-requiring-type-checking",
+        "plugin:import/typescript",
+        "prettier",
+        "prettier/@typescript-eslint"
+      ],
+      "rules": {
+        "@typescript-eslint/explicit-module-boundary-types": "off",
+        "eslint-comments/disable-enable-pair": [
+          "error",
+          { "allowWholeFile": true }
+        ],
+        "eslint-comments/no-unused-disable": "error",
+        "import/order": [
+          "error",
+          { "newlines-between": "always", "alphabetize": { "order": "asc" } }
+        ],
+        "sort-imports": [
+          "error",
+          { "ignoreDeclarationSort": true, "ignoreCase": true }
+        ]
+      }
+    }
+  ]
+}
