+++ .prettierignore
@@ -1,0 +1,5 @@
+# package.json is formatted by package managers, so we ignore it here
+package.json
+package-lock.json
+.nyc_output
+build
