+++ package.json
@@ -2,40 +2,83 @@
   "name": "ahv13-validator",
   "version": "1.5.1",
   "description": "Validates the AHV 13 Number (swiss social security number)",
-  "main": "index.js",
-  "scripts": {
-    "test": "grunt test",
-    "build": "grunt"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/longstone/js-ahv13.git"
-  },
+  "main": "build/main/index.js",
+  "typings": "build/main/index.d.ts",
+  "module": "build/module/index.js",
+  "repository": "https://github.com/groovytron/ahv13-validator",
+  "license": "ISC",
   "keywords": [
     "AHV13",
     "SSN"
   ],
   "author": "longstone <contact@longstone.ch>",
-  "license": "ISC",
+  "scripts": {
+    "build": "run-p build:*",
+    "build:main": "tsc -p tsconfig.json",
+    "build:module": "tsc -p tsconfig.module.json",
+    "fix": "run-s fix:*",
+    "fix:prettier": "prettier \"src/**/*.ts\" --write",
+    "fix:lint": "eslint src --ext .ts --fix",
+    "fix:markdown": "prettier \"**/*.md\" --write",
+    "fix:json": "prettier \"**/*.json\" --write",
+    "test": "run-s build test:*",
+    "test:lint": "eslint src --ext .ts",
+    "test:prettier": "prettier \"src/**/*.ts\" --list-different",
+    "test:unit": "jest",
+    "check-cli": "run-s test diff-integration-tests check-integration-tests",
+    "check-integration-tests": "run-s check-integration-test:*",
+    "diff-integration-tests": "mkdir -p diff && rm -rf diff/test && cp -r test diff/test && rm -rf diff/test/test-*/.git && cd diff && git init --quiet && git add -A && git commit --quiet --no-verify --allow-empty -m 'WIP' && echo '\\n\\nCommitted most recent integration test output in the \"diff\" directory. Review the changes with \"cd diff && git diff HEAD\" or your preferred git diff viewer.'",
+    "watch:build": "tsc -p tsconfig.json -w",
+    "watch:test": " jest --watch",
+    "watch:doc": "typedoc --watch",
+    "cov": "run-s build test:unit cov:jest && open-cli coverage/index.html",
+    "cov:jest": "jest --coverage",
+    "cov:send": "run-s cov:lcov && codecov",
+    "cov:check": "jest --coverage",
+    "doc": "typedoc && open-cli build/docs/index.html",
+    "doc:publish": "gh-pages -m \"[ci skip] Updates\" -d build/docs",
+    "version": "standard-version",
+    "reset-hard": "git clean -dfx && git reset --hard && npm i",
+    "prepare-release": "run-s reset-hard test cov:check doc:html version doc:publish"
+  },
+  "engines": {
+    "node": ">=10"
+  },
   "devDependencies": {
-    "@babel/cli": "^7.17.0",
-    "@babel/core": "^7.17.0",
-    "@babel/preset-env": "^7.16.11",
-    "babel-loader": "^8.2.3",
-    "babel-preset-minify": "^0.5.1",
-    "cross-env": "^7.0.3",
-    "file-loader": "^6.2.0",
-    "grunt": "^1.3.0",
-    "grunt-babel": "^8.0.0",
-    "grunt-check-dependencies": "^1.0.0",
-    "grunt-cli": "^1.3.2",
-    "grunt-contrib-watch": "^1.1.0",
-    "grunt-eslint": "^24.0.0",
-    "grunt-mocha-test": "0.13.3",
-    "grunt-nyc-mocha": "1.0.0",
-    "load-grunt-tasks": "^5.1.0",
-    "mocha": "9.2.0",
-    "nyc": "^15.1.0",
-    "webpack": "^5.68.0"
+    "@istanbuljs/nyc-config-typescript": "^1.0.1",
+    "@types/jest": "^27.4.0",
+    "@typescript-eslint/eslint-plugin": "^5.12.0",
+    "@typescript-eslint/parser": "^5.12.0",
+    "codecov": "^3.5.0",
+    "eslint": "^7.8.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-eslint-comments": "^3.2.0",
+    "eslint-plugin-import": "^2.22.0",
+    "gh-pages": "^3.1.0",
+    "jest": "^27.5.1",
+    "npm-run-all": "^4.1.5",
+    "open-cli": "^6.0.1",
+    "prettier": "^2.1.1",
+    "standard-version": "^9.0.0",
+    "ts-jest": "^27.1.3",
+    "ts-node": "^9.0.0",
+    "typedoc": "^0.22.0",
+    "typescript": "^4.5.5"
+  },
+  "files": [
+    "build/main",
+    "build/module",
+    "!**/*.spec.*",
+    "!**/*.json",
+    "README.md"
+  ],
+  "prettier": {
+    "singleQuote": true
+  },
+  "nyc": {
+    "extends": "@istanbuljs/nyc-config-typescript",
+    "exclude": [
+      "**/*.spec.js"
+    ]
   }
 }
